% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{nhsbsa_list_datasets}
\alias{nhsbsa_list_datasets}
\title{List Available Datasets}
\usage{
nhsbsa_list_datasets()
}
\value{
A character vector of dataset IDs available in the portal
}
\description{
Retrieves a list of all available datasets from the NHSBSA Open Data Portal.
}
\examples{
\dontrun{
# List all available datasets
datasets <- nhsbsa_list_datasets()
print(datasets)
}
}
