% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhsbsa-package.R
\docType{package}
\name{nhsbsa-package}
\alias{nhsbsa}
\alias{nhsbsa-package}
\title{nhsbsa: Client for the NHS Business Services Authority Open Data Portal}
\description{
A low-level client for the NHS Business Services Authority (NHSBSA) Open Data Portal \url{https://opendata.nhsbsa.net}, a 'CKAN' data catalogue. Provides thin wrappers around the portal's API actions for listing datasets, retrieving metadata, querying the datastore and downloading resource files. Results are returned as plain data (tibbles and lists) for the caller to interpret.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rmgpanw/nhsbsa}
  \item \url{https://rmgpanw.github.io/nhsbsa/}
  \item Report bugs at \url{https://github.com/rmgpanw/nhsbsa/issues}
}

}
\author{
\strong{Maintainer}: Alasdair Warwick \email{alasdair.warwick06@gmail.com} (\href{https://orcid.org/0000-0002-0800-2890}{ORCID})

}
\keyword{internal}
