% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhsbsa-package.R
\docType{package}
\name{nhsbsa-package}
\alias{nhsbsa}
\alias{nhsbsa-package}
\title{nhsbsa: Client for the NHS Business Services Authority Open Data Portal}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A low-level client for the National Health Service Business Services Authority (NHSBSA) Open Data Portal \url{https://opendata.nhsbsa.net}, a 'CKAN' data catalogue. Provides thin wrappers around the portal's API actions for listing datasets, retrieving metadata, querying the datastore and downloading resource files. Results are returned as plain data (tibbles and lists) for the caller to interpret.
}
\section{Further reading}{

The portal is a standard CKAN catalogue. For background on the API this
package wraps, see:
\itemize{
\item The portal's API page:
\url{https://opendata.nhsbsa.net/pages/api}
\item The CKAN Action API reference:
\url{https://docs.ckan.org/en/latest/api/}
}

\code{vignette("nhsbsa")} explains how the package maps onto what you see on the
portal website, and shows the different ways to query data.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rmgpanw/nhsbsa}
  \item \url{https://rmgpanw.github.io/nhsbsa/}
  \item Report bugs at \url{https://github.com/rmgpanw/nhsbsa/issues}
}

}
\author{
\strong{Maintainer}: Alasdair Warwick \email{alasdair.warwick06@gmail.com} (\href{https://orcid.org/0000-0002-0800-2890}{ORCID}) [copyright holder]

}
\keyword{internal}
